scripts/requirements.txt
matplotlib >= 3.5
msgpack >= 1.0
numpy >= 1.22
pyserial >= 3.5
rich >= 12.3